agent/pyproject.toml
[project]
name = "capstone-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "assistant-stream>=0.0.30",
    "azure-cosmos>=4.5.0",
    "fastapi[standard]>=0.117.1",
    "gunicorn>=23.0.0",
    "httpx>=0.27.2",
    "langchain>=1.0.0a9",
    "langchain-google-genai>=3.0.0",
    "langchain-openai>=0.3.33",
    "langgraph-api>=0.4.43",
    "langgraph-checkpoint-mongodb>=0.2.1",
    "langgraph-cli[inem]>=0.4.2",
    "langgraph[inem]>=0.6.7",
    "msgpack>=1.1.2",
    "pydantic-settings>=2.11.0",
    "pyjwt[crypto]>=2.10.1",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.1",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
